(refactor) rename strat commands
