+++ lib/wss/cmds/start_strategy.js
