+++ lib/wss/events/on_message.js
@@ -4,16 +4,16 @@ const _isFunction = require('lodash/isFunction')
 const debug = require('debug')('bfx:hf:algo-server:wss:events:on-message')
 const getAllAOs = require('../cmds/get_all_aos')
 const submitAO = require('../cmds/submit_ao')
-const submitStrategy = require('../cmds/submit_strategy')
+const startStrategy = require('../cmds/start_strategy')
 const stopAO = require('../cmds/stop_ao')
 const stopStrategy = require('../cmds/stop_strategy')
 
 const cmdMap = {
   'get.aos': getAllAOs,
   'submit.ao': submitAO,
-  'submit.strategy': submitStrategy,
   'stop.ao': stopAO,
-  'stop.strategy': stopStrategy
+  'strategy.start': startStrategy,
+  'strategy.stop': stopStrategy,
 }
 
 module.exports = async (asState, ws, msgJSON = '') => {
